DHCPREQUEST now answered when no IP is set
We found that the VMs were sending DHCPREQUESTs and they were not
answered by midolman. MM reported an error when the request didn't
include an IP. Now those requests are answered with the previous
IP.

Also changed the warning messages to debug.

Tested on a real system. No unit or functional tests run.

Change-Id: I1d525709e7d522f8e14059aea9d5890bf17858f5
Signed-off-by: Abel Navarro <abel@midokura.com>