Bridge handles ARPs; expose mac and arp table.
This commit exposes the bridges mac table via the API. Its entries may
be listed and it's now possible to add/remove persistent entries. The
persistent entries may be set before ports on the bridge use the mac
so that Flooding is avoided.

This commit also uses static dhcp host assignments to populate a new
per-bridge IP4-to-MAC map. When ARP replying is enabled, the bridge
uses the IP4-to-MAC map to reply to broadcast ARP requests. This
avoids flooding when the IP to MAC mapping is known by the layer
that configures the virtual topology.

The commit also exposes the Bridge's ARP table via the REST API.

Bridges that pre-date the ARP feature will not have the IP4MacMap
sub-directory. We need to make sure that:
- new static host DHCP assignments can be created. i.e. they shouldn't
trigger ZK errors because the attempt to add the IP4-Mac mapping fails
to find the ARP table directory.
- the bridge is correctly loaded for simulation. i.e. the attempt to
initialize a ReplicatedMap based on the ARP table directory should not
encounter a NoNodeException.
- the bridge may be deleted. i.e. the deletion routine will not attempt
to delete a non-existent ARP table directory (and then fail to delete
the bridge).

We handle the first two cases by creating the ARP table directory if
it does not already exist. We handle the third case by checking whether
the directory exists before attempting to prepare a recursive delete
on its contents.

Change-Id: Iab980f6009f4c689c95a263bd076bce763389e0d
Signed-off-by: gdecandia <gdecandia@midokura.jp>