Remove the actor code from PWA.
We want to get rid of akka to improve the performance for the PWA as it is
only used once in a simulation, also, it only receives one message during
its lifetime so this actor is easily converted to a normal class with methods.

The actor inherited some traits which were meant to be used by an actor
implementation. Refactored those to use passed in actor system & context
instead.

Change-Id: If40d82731f9f93b3d7b5b081a7d8c20ed961e031
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>
Signed-off-by: Marc de Palol<phlegias@gmail.com>