Correct ARP logic. ARP even if cache entry expired.
ArpCacheEntries that have expired should be removed automatically.
However, we had a bug (#551) where old entries were not cleaned up
properly in the underlying map. That confused our algorithm because
we don't send ARP requests when the entry is expired (we interpret
that to mean that there are no more retries left). The workaround
is to replace the expired entry with a new entry (with null MAC)
before starting the ARP-retry loop.

Change-Id: I40ff4c9fa2e5871e1cff4b441e648aea48f23e56
Signed-off-by: gdecandia <gdecandia@midokura.jp>