Fix PING to FloatingIp in router's far port
This fix simply takes care of applying some of the post-chain
transformations to the generated packets by the router. This will allow
the SNAT to be applied on the source ip sent on the ICMP Echo Reply.

The fix is somehow dirty because it involves applying a subset of
transformations inside the router itself.

The other alternative was to refactor all the packet processing
code in order to allow the Router to send a WildcardMatch to the
PacketWorkflowActor instead of just the generated Eth. frame. Although
that would be cleaner, after discussing with Pino we agreed on this
solution because:

- The size of the change would be much bigger (higher risk + dev cost)
- Having the router port assigned a floating IP to itself is not a
  common use case.
- It is unlikely that more complex transformations will happen on post
  rule chains in this situation.

Issue: #547
Change-Id: I92763bebfd76826f45579a45e445b175bcc3cccd
Signed-off-by: Galo Navarro <galo@midokura.com>