BufferPool: keep buffers in an identity map (#576)
ByteBuffer implements hashCode() based on the buffer content
so they cannot be used as map keys to track them inside BufferPool.

The effect of this bug is that the BufferPool never reclaims buffers
when release() is invoked, because it doesn't find them in its pool.

The fix is to use an identity map instead.

Change-Id: I506c03c527e7765a3a6c43bb57955b003f1b8fee
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>