Expire fwd flows before ret flows
We want the fwd flow to expire earlier than the return flow so that the
cache key gets refreshed and the return flow can be later identified as
as such.

Upon installation of the fwd flow a cache entry is installed in
Cassandra to allow identifying the return flow. When a return packet
is simulated, MM will fetch this key, identify it as return and install
the corresponding flow.

However, this cache entry has a much smaller expiry time than the fwd
flow. This implies that the cache entry may disappear, even though the
fwd flow is still active. The return flow expiry time is also lower than
the fwd flow's. This implies that you may run into situations where a
ret. flow is being simulated, its corresponding fwd flow is active, but
the cache entry isn't present, thus making it impossible to identify the
ret. flow. This will result in DROPping the return flow.

By setting the fwd flow's expiration time to a value lower than the
ret. flow's, we ensure that fwd packets will keep triggeting refreshes
of the cache key before the situation above ensures.

Also, when a conntrack entry is not found in Cassandra, make the
resulting DROP temporary to allow faster recovery.

Issue: #577
Change-Id: I0bb282a3f3956093bea6b6290315a915464aa108
Signed-off-by: Galo Navarro <galo@midokura.com>