restricted grep range for DhcpInterfaceMtu test
The grep command in DhcpInterfaceMtu to find out the mtu of the interface
could fail with some outputs of ifconfig on osx.
Restricted range of grep to fix that.
This is only a temporary fix, eventually such external calls should be
mocked away.

Signed-off by: hugo benichi <hugo@midokura.com

Change-Id: I4a2a6247ae1d4e5e7854a522ed02191b728050a5