Improved handling of idle time-out expiration
When a wcflow has expired according to the lastUsedTime we have, before evicting it we
ask for an update of the lastUsedTime of the corresponding flows
to the kernel.
We make a decision about the wcflow expiration only when we receive the
fresh data for all the corresponding kernel flows.
Map of kernel flows -> wcflows was removed.

Change-Id: I979c7c16d47ca1bd9b5e41afb83edb1af950ea96
Signed-off-by: Rossella Sblendido <rossella@midokura.com>