Disable failing DeletePortTest
Looks like it was not fixed completely when enabling it. Since it was
never really working, I'm disabling it again.

A likely cause is dodgy handling of deleted ports. There are 3
expectations in the probe for a LocalPortActive = false, plus 2 deleted
flows. These do happen according to the logs. The problem is that the
PortZkManager explodes since the deleted port doesn't exist in ZK after
the deletion. But it looks like this is causing a NoNode exception that
ultimately breaks the test.

Issue: #605
Change-Id: I5095d0a1211a69b0d9878c76955ac5f568a348b1
Signed-off-by: Galo Navarro <galo@midokura.com>