WildcardFlow: create a pool and reuse objects
First, alter the WildcardFlow hierarchy so that the objects maintained
by the FlowController are reusable. WildcardFlowBuilder disappears,
a common trait for both classes is created (WildcardFlow), and a new
reusable class appears: ManagedWildcardFlow

A new ObjectPool collection is introduced to maintain a pool
of reference counted objects. Reference counting is implemented in a
trait that can be mixed into any class. This trait is mixed into
ManagedWildcardFlow.

The FlowController uses objects taken from the pool when it needs to
add a new WildcardFlow, copying the received flow's content on top of
the pooled object. Both the FlowController and the FlowManager update
the flow's reference count as needed, and the objects are put back
in the pool automatically by the ref-counting trait when the count
reaches zero.

The maximum number of wildcard flows becomes a mandatory configuration
parameter, it can no longer be unbounded.

Change-Id: I60833e44f98cc02d52c1da4b54ff78eb385f6dba
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>