Add new vlan aware device, new ports, ZK classes
Vlan Aware bridges have 2 exterior ports ("trunks") and a number of
interior ports that can be connected to virtual bridges, each being
assigned a single VLAN ID.

For frames received on any trunk, the VlanAwareBridge will:
- If it's a BPDU, forward it to all the other trunks
- If the frame has a VLAN tag, find the interior port that has the
  corresponding VLAN ID. If there is none, the frame will be dropped.
  Otherwise, it'll add a POP VLAN action and put it on the interior
  port.

For frames received on any interior port, it will add a PUSH VLAN action
with the VLAN ID corresponding to that port and egress the frame from
*all* the trunk ports (under the assumption that as a result of STP
operating among the switches connected to the trunks only one of these
will be active).

Most of the changes are relatively self contained and only add new API
functionality and custom logic in the simulation code. There is also a
change to the treatment of ToPortSetAction. Before, this action was
always assumed to happen on Bridges, but now it may also happen on
VlanAwareBridge (when outputting a frame from the virtual network to all
the trunks).

Change-Id: Iefc2f7bda465dc725dbe911b06cbfe590492c738
Signed-off-by: Galo Navarro <galo@midokura.com>
Signed-off-by: Rossella Sblendido <rossella@midokura.com>