Fixed a problem with dlType in DTO
When fixing issue #537, unit tests didn't exercise DTOs and the one
regarding the dlType failed in the functional test.

This is a fast fix which forces the user of setDlType to use Integers
instead of Shorts. The problem was that when converting from Short to
Integer in that function, Java complained about precision loss.

Reference: issue #537

Change-Id: If22947f4f707b0d727b60e30839fe24a6a09e81c
Signed-off-by: Abel Navarro <abel@midokura.com>