RoutingManagerActor to use tell, not expiringAsk
This change is actually reverting the behavior of RoutingManagerActor
when asking for a port to the VirtualTopologyActor. Previously it was
using a tell to send a message and now it was using expiringAsk.

The way RMA communicates with VTA does not need a future, just a simple
message. VTA will send an ExteriorRouterPort message back, not a future.

It has been briefly tried at the QA stack with positive results.

Reference: issue #630

Change-Id: Icc79059e5167b813a133b1b1418efbd4f3079781
Signed-off-by: Abel Navarro <abel@midokura.com>