FlowTranslatingActor: chain tags for translated flows
The FlowTranslatingActor (doing work formerly performed by the
DatapathController) was not adding chain tags after applying
the outbound filters for ports belonging to a port set. This
was the cause of issue #621.

For the same reason, the PacketWorkflowActor will now add those
same tags when a tunneled packet is received addressed to a
port set.

A unit test is included.

Fixes #621.

Change-Id: I9cbec0680a32fbe809f17e8f34bebf4bd44a0758
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>