Call IPSubnet.containsAddress only for same IPvX.
The Condition class uses IPSubnet for the source/destination network
prefix regardless of the IP version. Similarly, WildcardMatch uses
IPAddr for the match's source/destination regardless of IP version.
Prior to this patch, Condition.matchIP was not checking whether the
condition's IPSubnet was the same IP version as the match's IPAddr
before calling IPSubnet.containsAddress(IPAddr). This resulted in
a cast exception.

Fixes #624.

Change-Id: I11c824eae7b2ed019b3b4982ae096ac69565d23b
Signed-off-by: gdecandia <gdecandia@midokura.jp>