FlowController: add flow invalidation history
Introducing a rolling EventHistory class used by the FlowController
to maintain a history of flow invalidations. This history is used to
prevent invalidation races with flows installed by the PWA. The PWA
records the last known invalidation when it's initialized, and sends
it to the FC when it's time to add a flow.

Fixes #470.

Change-Id: I80b7616a0323391d1b04cbfc5eeefb0b1a930070
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>