Fixes for BgpTest functional test
There were several problems with BgpTest, mainly due to wrong paths used
for bgpd and zebra. This was due to the environment used to first
develop the test.

The binaries for bgpd and zebra are now set as final strings and we do
check if they exists, so in case the environment is missing those, the
test will fail much earlier. We do the same with the config files,
needed for this test.

A new config file (peer.zebra.conf) was added.

The BgpTest now uses a TestWatcher to report test start/finish and if it
succeeded/failed. It is hard to know when those events happen just
looking at the logs. This is a first step into moving this new printouts
to a base class, because all tests would benefit from this.

Change-Id: I3c089c42017fccfd44447dcdd28efd4c4c08a684
Signed-off-by: Abel Navarro <abel@midokura.com>