Add vlan-aware capabilities to old Bridge
Bridges can now react to vlan-tagged frames in a way consistent with the
legacy Virtual Aware Bridge, but also adding existing functionality for
MAC learning. There is some tuning in frame management, but it's
functional.

There is a slight change in functionality related to multicast frames
coming from the trunk that don't have a VLAN ID set. In the legacy VAB
case, the frame would be immediately dropped (unless it was a BPDU).
Here we simply treat its as an ordinary one and send it to the port set.
This is specially relevant with unknown dst. IPs where ARPs are
generated. Say a packet with no vlan-id comes from a trunk addressed to
an unknown ip. In the enhanced Bridge, the resulting ARP will be sent to
all trunks whereas in the VAB it would have been dropped.

Also includes a refactor to the VAB unit test to allow Bridge testing

This refactors the test and arranges in a way that provides the basic
helpers, test cases, etc. just relying on entry points to a topology
(the trunks, the VM's interfaces, etc.). Then extend with two cases, one
that builds a topology with the existing VlanAwareBridge, another that
puts a Bridge.

Subcommit for multicast from vlan-aware bridge (ForkAction):

This covers the previous functionality that allowed several
Coordinator.Actions in the Fork, but also allows more than two, and
incorporates a new type of Coordinator.Action that wraps a FlowAction.
Upon processing this action, the coordinator will generate an
AddWildcardFlow that later will be merged with the rest of
Simulation.Actions resulting from the rest of actions.

Hopefully the previous paragraph is enough explanation of why Guillermo
and I decided to rename SimulationAction to SimulationResult.

The use case this is required for is vlan-tagged frames with a multicast
destination from trunks into vlan-aware bridges. An example is an ARP
request. The bridge will now send this frame to both the interior port
tagged with such Vlan id, but also to its port set. Here we need to take
into account that the frame sent to the interior port must have the
vlan-id popped *before* the bridge on the other side of the port
simulates it. Whereas the frame that goes to the vlan-aware bridge's
portset needs to keep the vlan id. So, the resulting actions are:
- Send to port set (a coordinator action)
- Pop vlan (a coordinator action that wraps a flow action)
- Send to port (a coordinator action)

The Coordinator ensures that each of these happen sequentially, injects
the AddVirtualWildcardFlow with the POP vlan, and modifies the matches
accordingly.

We are not quite happy with the fact that the Coordinator has to
translate a FlowAction into the packet context, but at this point making
something more sophisticated would increase the scope quite a bit, but
improving this piece would be a desirable followiup.

The legacy VAB has not been changed accordingly and still will *not*
send multicasts from a trunk to the other trunks (following the original
spec from Sakura). The vlan-aware Bridge is thus a generalisation of the
VAB.

Unit tests are passing, the functional tests using a topology with the
Bridge fail due to the change in expectations.

Documentation includes removal of references to the old VlanAwareBridge
and its ports since this becomes deprecated as of now.

Change-Id: I933945d066c30f729c97407dbe25cc032d692ad4
Signed-off-by: Galo Navarro <galo@midokura.com>