BGP: fix handling of LocalPortActive
Handling of LocalPortActive(false) in the RoutingManagerActor was
broken, it was stopping the RoutingHandler actor (without giving it
a chance to clean up) and not enabling it again if the port became
active again. Also, stopping the actor is the wrong policy, because
the ZK managers don't allow unsubscription/resubscription, impeding
a second actor from taking over from the original one.

With this change:

 * When a port becomes inactive, the RoutingManagerActor will forward
   the information to the relevant RoutingHandler.
 * The RoutingHandler will enter the 'Disabled' state, delete all the
   discovered routes and stop bgpd.
 * Similarly, when the port becomes active again, the RoutingManagerActor
   will tell the RoutingHandler, which will restart bgpd and rediscover the
   routes, if the peer is still announcing them.
 * The RoutingManagerActor will no longer stop the RoutingHandler. However,
   the postStop() event in the RoutingHandler will delete all the
   discovered routes and try to stop bgpd.

Fixes #674 and includes 2 functional tests.

Change-Id: I8dea78e0e74f4885eecb83e09d890bd99c20e725
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>