Start a new top-level "Build" script.
This currently verifies that we have the correct version of Java
available and that Maven is installed, and then kicks off a build (not
cleaning first--we need to add a -c option) and run of the tests,
skipping the functional tests.

In the long run, it should by default run all developer tests that
run quickly and are guaranteed to be passing, producing no output
(unless asked to be verbose) when successful and just the information a
developer needs to identify the errors, when not.

Change-Id: Ifbd9c04815142a46769034d81952ab2127971fc7