Use Range objects for port ranges
This way we can avoid interpreting (0, 0) as no range, which was causing
problems detailed in the issue (e.g.: an ICMP rule with 0, 0 is
interpreted as "all ICMP" when in reality the tpSrc and tpDst are the
ICMP type and code, therefore letting ICMP requests pass when the
intention was to let only ICMP replies).

This changeset includes a new DtoRange that can be used with other types
of contents, and a Range object in the util package.

Some test cases became irrelevant since the Range validates that start
is less or equal than end.

Note that these changes will need to be reflected in the Python client.

Issue: #495
Change-Id: I5d08b5ac85a58e3b0647823d5a0ba8ca1a7557ea
Signed-off-by: Galo Navarro <galo@midokura.com>