Stabilize TestCases: Ping,FlowManagementForPortSet
Some more changes to the PingTestCase to clear probes etc. and try to
prevent races.

Also, fix to the flow managemet one. I suspect that this may be related
to the order of fishes because the error seen in Jenkins is not that it
failed to retrieve a message, but that the contents were not those
expected. It now just tries to get all the expected messages, and
accepts any of them. At the end, it expects to have 0 pending.

Issue: #618, #683
Change-Id: I7e18d4dcf7f6120d1bd79171c6f40cec28d24a6d
Signed-off-by: Galo Navarro <galo@midokura.com>