pom.xml: add XML version tag and tweak formatting of project tag.
The XML version tag was missing; this one was taken from the tweaks
that IntelliJ Idea makes to this file.

The project tag is merely reformatted by Idea; we follow their
formatting to reduce the number of changes it makes, in preperation for
figuring out why it's making its other changes, and if perhaps the code
we have is simply not needed.