.gitignore: stop ignoring .swp files.
These are typically generated by vim (and perhaps other editors). Given
the wide variety of these, it's best to set your editor not to use swap
files, to put the swap files in a different place, add the patterns to
your personal core.excludesfile setting in $HOME/.gitconfig, or to add
the patterns to .git/info/exclude in your copy of the repo.

For vim, "set noswapfile" will turn off swapfiles.