Setup: revert ZK directory structure creation to old code, with fixes.
Rename midolman.Setup.createZkDirectoryStructure() to
ensureZkDirectoryStructureExists to clarify its intent. We don't care if
the ZK nodes exist or not before we started; all we want is to ensure
that they're there before moving on to other operations using them.

Thus, there's no point to checking for existence at some arbitrary time
before we call rootDir.add, since it could be created between that
time and the rootDir.add call anyway. If the node already exists, the
rootDir.add call is highly likely to be as cheap as just a check for
existence.

Logging something if the directory did exist before the call, but not if
it didn't, was misleading since we don't really care about that.

Change-Id: I6d93692079db108f52b94f6a3d659c440a0cc1e7