Build: show error messages when checking for presence of Maven.
The `mvn --verbose` command, when successful, produces output only to
stdout. So we keep redirecting that to /dev/null, but stop sending
stderr there so that we can see a bit more information about what went
wrong when we can't run Maven.

This was prompted by a Maven failure that was not due to it not being
installed, but due to having non-world-readable configuration files in
/etc.

Change-Id: I3f208023d83676c6d30156770c9b60d56decee35