Fix CacheModule and ConditionSet injection
Guice was exploding when trying to inject the cache modules after
changes made in rev. 963485076fadba1b99a7b4487c508e3da6097302, this
patch adds the relevant annotations in CacheModule and corrects the
bindings in MidolmanActorsModule. Also required some changes in the
MockCache binding used in Unit Tests.

Re. ConditionSet, the injection was not properly configured for
EmbeddedMidolman, which was killing functional tests.

Issue: #693
Change-Id: I7e00cb8864c5455bcec3024b920baa3acb667412
Signed-off-by: Galo Navarro <galo@midokura.com>