Add remove by (k,v) to ReplicatedMap
This solves a risk of inconsistency in the Replicated Map:

BridgeManager -  Forgetting mac-port entry 4a:ee:59:0f:7b:16
    22b1dff6-8217-4e7b-8c45-0aa2eb611ab9
BridgeBuilderImpl$MacTableNotifyCallBack -  Mac-Port mapping for MAC
    4a:ee:59:0f:7b:16 was updated from f2b2e98e-c870-4719-ab95-67400d9c0076 to null

Note how the value deleted (second line) wasn't the one that was actually
requested for deletion.

This example happened after port f2 has been learned for mac 4a, replacing the
old port 22. As a result of this:
- Flows for 22 are invalidated by tag which gradually brings the count of flows
  for mac-port down to 0. At this point a removal for port 22 is scheduled.
- When the scheduled removal kicks in, the MacLearningTableImpl executes
  the removal

The problem was that MacLearningTableImpl.remove is based on both mac and port,
but it relies on ReplicatedMap.removeIfOwner which isn't. So it just wipes off
whatever value the key has.

The second log message above with the update callback is actually accurate, the
key's currenct val was removed from the map and the value passed from f2 (that
is: the value that replaced the old 22) to null (removed).

The problem with this is that the facts inferred by the MacTableNotifyCallback
are wrong and it won't invalidate the right flows, causing much grief.

The solution has been to add a new method in the ReplicatedMap that also
considers the value that the clien is trying to remove. If the value doesn't
match, it won't delete the entry.

Including some javadoc in the ReplicatedMap.

Issue #690
Refs #684
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Id4257c9925e760bda75b2f4061575793b233e013