Do callbacks if if flow skipped or duplicate
If upon flow installation we see a Netlink error reporting that "file
exists while adding flow", or if a flow installation is skipped because
it's obsolete, we still need to perform the callbacks. This was
generating inconsistent flow counts in the mac-learning table which in
turn prevented the right invalidations from being fired.

Issue #690
Refs #684

Change-Id: Ia4f7fb0e22db86d41da6e073b194ac090cadddce
Signed-off-by: Galo Navarro <galo@midokura.com>