Tag L2 unicast flows with srcMac, inPort in Bridge
This is necessary in the event that the mac changes a port (e.g::
failover scenario with L2GW). Say a MAC moves from portA to portB, if
any flow count in the mac-learning table for the pair mac-portA remains,
when the failback to portA happens the bridge manager will find a flow
already and not realize that it's a port change so it won't learn the
new port, effectively cutting traffic.

Issue #690
Refs #684

Change-Id: Ib8ce52b90dc55201c1137ca115a3597b92a3fcdf
Signed-off-by: Galo Navarro <galo@midokura.com>