Prevent wmatch field set corruption
The FlowController was using the field set provided by WildcardMatch as
keys in the WildcardMatch tables, but this set was a reference to the
private data structure. When the first WildcardMatch that provided a
given Set of fields was returned to the pool, the internal set is
emptied, thus corrupting the key in the FlowController's table.

After discussing with Guillermo, we decided for the solution above.
Although it would be much safer to make defensive copies on
WildcardMatch.getUsedFields, this would cost us a lot of unnecessary
copies. More or less the same if we recreated the Set upon WM.clear(),
instead of clearing it. I'm adding due WARNINGS everywhere anyway.

This also includes a fix in the WMatch related to vlanIds, which were
not being managed properly.

The test case was done while debugging the inconsistent entries in the
WM tables, in the end they are not really related, but since I had
already written it I guess it doesn't hurt to get it in.

Issue #690
Ref #686
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ic284f163c4861328df31bea49cfcde9d86d04523