Pass-through fragmented packets if no L4 match
This changeset will allow fragmented packets to pass through as long as
no L4 fields are accessed in the WcMatch by any of the devices traversed
during the simulation.

From this change, devs must be careful when accessing fields in the
WcMatch since this may have side effects when deciding whether to let
fragments pass or not. A method has been added to enable/disable
tracking of fields used in order to allow peeking into values (this is
necessary for example when performing diffs).

Change-Id: I9f72e3a939cb1251d52c1920d9fff4c4b8904a9a
Signed-off-by: Galo Navarro <galo@midokura.com>