Untie Midolman layer from NetlinkChannel
The goal of this patch is to untie the midolman layer from direct
imports of NetlinkChannelImpl and MockNetlinkChannel. This will
allow to remove jdk-boostrap dependency in midolman and odp
layers at compile time, keep it only at runtime and ease the
transition to sbt-zinc based build system.

The main change for this is to have the OvsDatapathConnection
constructor take a Netlink address in place of an already opened
NetlinkChannel. Instead, the OvsDatapathConnection makes its own
channel by accessing the NetlinkSelectorProviver internally and
connecting it to the provided address.

Additionally the NetlinkSelectorProvider can be accessed in a
safe way via the Netlink class.

Furthermore NetlinkChannelImpl and its mock brother are now
package private. Clients (such as OvsDatapathConnection) only sees
NetlinkChannel typed objects.

Finally, this also addresses GH #137 and merge the two NetlinkClient
classes into a single DatapathClient class (name reflecting the fact
it creates a working OvsDatapathConnection instance).

Issue #137

Change-Id: I38cff06afed5518b8a92d74a46a5982567899f30
Signed-off-by: hugobenichi <hugo@midokura.com>