Refactoring api port classes, without changing api itself.
Changing the api's Port classes to be able to support
unplugged ports in later commits. The new api will take in
BridgePort and RouterPort, and the Interior*Port and
Exterior*Port classes will be deprecated. I've moved the fields
and methods to their superclasses, but Interior/Exterior*Port
classes still exist for the current api.
For details on why this is happening see:
https://sites.google.com/a/midokura.com/wiki/
   midonet/port-types-and-states
********Test Deletions********
Since apparently all vlan*Port and vlan*Bridge classes are now
obsolete, I removed a bunch of test code from TestPort and
ClientTest. Please review to see if any of this was actually
needed.

Change-Id: I12619ad740f4c4d6879bfb0b724709e18abc6db6