risky block of code was evaluated before try catch
Some method which was trying to wrap a block of code in a try catch
statement had a bug in it because the block of code was passed as a
call-by-value argument and evaluated at the method invocation. Fixed
by changing to call-by-name semantics. Found while doing code reading
with Brandon.

Change-Id: I16e0a76241dc43c020a5edd22d785b4904ae3999
Signed-off-by: hugobenichi <hugo@midokura.com>