Set bridge peer port id also if empty
The builder object started out with vlanBridgePeerPortId as None, and
sets it to something when it sees a new port. As it never gets set back
to null even if the port is removed. This changeset makes sure that if
the peer port disappears, the field will be reset to empty.

Issue 744
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I2d68ea42a8de392e20386fda092a4b5b5aebef95