Set bridge peer port id also if empty
The builder object started out with vlanBridgePeerPortId as None, and
sets it to something when it sees a new port. As it never gets set back
to null even if the port is removed. This changeset makes sure that if
the peer port disappears, the field will be reset to empty.

Change-Id: I490396a7d88aec395df70c304b7adde1486fdde7
Signed-off-by: Galo Navarro <galo@midokura.com>