Refactoring ports classes to remove interior/exterior types.
Continuing with making changes described here -
https://sites.google.com/a/midokura.com/wiki/midonet/
port-types-and-states
This commit removes interior/exterior ports from the cluster
classes, simulation port classes, config classes, and
java-client.
This changes the serialization as described in the
wiki - introduces a "v1ApiType" for backwards compatibility.
There is a new api using the new media types.
Tests are fixed to be compatible. Some VlanBridge tests
are removed. Vlan*BridgePorts are deleted.
Tunnel keys are now created for all ports, even though they're
only needed for exterior ports.
The bgp folder is created for router ports at creation time,
and deleted at port deletion. If a router port becomes
an interior port that folder just won't be used.
One significant change is that a router's routing table is not
updated in zk when a new port is created. Instead it is updated
when the port becomes linked. This means that if you create a
new port, do not link it yet, and add a route for it, the route
will not take effect until the port is actually linked. On
unlink, the routes are removed from the routing table.
The java client now only supports the v2 api, no backwards
compatibility, although the v1 api is still available for
use with an old client. Since all the tests use the client,
the v1 api has fewer (no) tests.
Includes a small bug fix:
--Tunnel keys are now deleted when both exterior bridges
  and routers are deleted. Previously they were only deleted
  when an exterior bridge was deleted.

Change-Id: Ibf2b183d0352001017de2abf2c5dca9af8846702