Improve future management in FlowTranslator
This was hiding errors during flow translation phases that hid various
problems. One in particular was that it seemed to let zombie cookies
associated to a match which turned the node into a black whole for all
packets with the same match, forever.

Also minor stylistic fixes and refactors to make scala code nicer and
safer.

Fix MN-204
Ref MN-76
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I4fe9361c72d707f8e57a5ee7a7b76ac67a182d95