Make vlan-unaware bridges work with vlan traffic
Vlan unaware bridges were failing to process vlan-tagged traffic because
the per-vlan mac table only gets a partition for a given vlan if there
is an interior port with that specific vlan. As a result, when a VUB
(with no vlan-tagged ports) received vlan-tagged traffic it failed to
do mac-learning properly.

This changeset changes the behaviour of the bridge. If it doesn't have
any vlan-tagged ports, then its mac-learning table will work as a
single, generic table, not partitioning by vlan.

Issue MN-282
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: I10cb919dd4c731d9d5b32740b3d9546b0069283d