Netlink: batch netlink messages processing
We introduce a new NetlinkCallbackDispatcher actor. This actor will install
a new "BatchCollector" in the datapath layer. This is an interface
that allows getting Runnable instances in batches. That actor's implememtation
of this interface will send the whole batch in one Akka message to the actor
for processing. Thus removing their execution from the netlink layer.

To make this possible, the netlink layer is thusly modified to call the
translation function on netlink buffers right away but create with a Runanble
that will invoke the real request callback when the batching dispatcher
decides to do so. A default dummy implementation is included in the datapath
to make the addition of this hook optional.

To batch incoming packets, we make the odp connection layer take a
BatchCollector instead of a callback to collect packets. The
DeduplicationActor implements the BatchingCollector by accumulating
packets in a list and sending the list as a whole to itself.

AbstractNetlinkConnection will invoke the overrideable endBatch() method at
the end of each select loop iteration. ODP connections override it to
notify their BatchCollector<Packet> instance.

Change-Id: I3611a18f375475cdb80735701a53ef6a6027c224
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>