packets: never fail on deserialize()
Fallback to using Data as payload for a packet when deserialize() fails on
what should be the rightful class of a given payload. This allows malformed
packets that should get through (malformed layers are not looked at by
midolman) to do so, instead of crashin on Ethernet.deserialize().

Fixes MN-68

Change-Id: Ic2c5ba21745d603cca21dde4411c7dfc636ff520
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>