clean DatapathController actor Messages types
Removed obsolete DatapathPort types removed in ovs 1.10. Changed the
typing of messages handled by the DPA related to port operations.
This makes the pattern matching much clearer to detect operation
success or failure. This also allows to eliminate sendOpReply() by
having OperationRequest generates their own Reply msg. The Reply msg
is made so that it carries the original Request msg.

Using this own Reply msg creation solution, it is easy to reschedule
port operations by resending the original request to the DPA without
directly calling functions createDatapathPort or deleteDapathPort
which are not completely thread safe and which are internal to the
DPA. This is applied to GreTunnelPort creaiton failure.

Fixes MN-300

Change-Id: I0105fcfd8ef9ae5b35f34c936ce3bb2fcbff9b91
Signed-off-by: hugobenichi <hugo@midokura.com>