MN-191: Do not throw an exception if setting a null value.
When setVifId or setPeerId is called on the wrong type of port,
do not throw a runtime exception if the value being set is null
anyway. This patch unfortunately fixes the two methods in
inconsistent ways. This is because in the following patch which
will be merged soon (https://reviews.midonet.org:8443/#/c/1022/),
it's done this way so its better to not have merge conflicts.
This can be cleaned up afterwards so it's not too bad.

Fixes MN-191

Change-Id: I465d36019f65d9ff11c9608df50ffaa2141cb013