Fix ICMP UNREACHABLE src mac from Router
The src mac in the packet's match was being modified too early with the
routers output port's mac. But this implied that if later in the
simulation the router decided to send an ICMP UNREACHABLE, the src mac
was not the original packet's but instead the routers out port.

This seems to have been there for a long time and hasn't been caught
because the tests were validating various fields in the ICMP error but
not the l2 src/dst. So I fixed the tests also.

Fixes #MN-612
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: Ic0862eda87d9988b07dd550131f98a99a2cc33c1