Coordinator: drop later fragments will a full match
Before fragment pass-thorugh was implemeted, Coordinator dropped
Later fragments silently, always. For that reason, the WildcardFlow
didn't need tags or callbacks, and it could match all Later fragments.

Now, Later fragments are dropped only when no L4 fields were looked at
in simulation. For this reason, the drop flows need to be as specific
as any other flow and carry the tags/callbacks.

Fixes: MN-621

Change-Id: I74a4420ecc7a4fed39f6d0c2f3ca762f16ce27b0
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>